refactor: move sensitive_columns property, _set_and_validate_sensitive_columns, and reveal_sensitive_columns into BaseCompare
